requirements.txt
gql==3.5.0
matplotlib==3.9.2
pytorch==2.4.1
requests==2.32.3
pandas==2.2.2
